+++ src/Symfony/Component/Mailer/Exception/UnsupportedHostException.php
@@ -21,27 +21,27 @@ class UnsupportedHostException extends LogicException
 {
     private const HOST_TO_PACKAGE_MAP = [
         'gmail' => [
-            'class' => Bridge\Google\Factory\GmailTransportFactory::class,
+            'class' => Bridge\Google\Transport\GmailTransportFactory::class,
             'package' => 'symfony/google-mailer',
         ],
         'mailgun' => [
-            'class' => Bridge\Mailgun\Factory\MailgunTransportFactory::class,
+            'class' => Bridge\Mailgun\Transport\MailgunTransportFactory::class,
             'package' => 'symfony/mailgun-mailer',
         ],
         'postmark' => [
-            'class' => Bridge\Postmark\Factory\PostmarkTransportFactory::class,
+            'class' => Bridge\Postmark\Transport\PostmarkTransportFactory::class,
             'package' => 'symfony/postmark-mailer',
         ],
         'sendgrid' => [
-            'class' => Bridge\Sendgrid\Factory\SendgridTransportFactory::class,
+            'class' => Bridge\Sendgrid\Transport\SendgridTransportFactory::class,
             'package' => 'symfony/sendgrid-mailer',
         ],
         'ses' => [
-            'class' => Bridge\Amazon\Factory\SesTransportFactory::class,
+            'class' => Bridge\Amazon\Transport\SesTransportFactory::class,
             'package' => 'symfony/amazon-mailer',
         ],
         'mandrill' => [
-            'class' => Bridge\Mailchimp\Factory\MandrillTransportFactory::class,
+            'class' => Bridge\Mailchimp\Transport\MandrillTransportFactory::class,
             'package' => 'symfony/mailchimp-mailer',
         ],
     ];