+++ src/Symfony/Component/Mailer/Bridge/Amazon/CHANGELOG.md
@@ -1,7 +1,15 @@
 CHANGELOG
 =========
 
+4.4.0
+-----
+
+ * [BC BREAK] Renamed and moved `Symfony\Component\Mailer\Bridge\Amazon\Http\Api\SesTransport`
+   to `Symfony\Component\Mailer\Bridge\Amazon\Transpor\SesApiTransport`, `Symfony\Component\Mailer\Bridge\Amazon\Http\SesTransport`
+   to `Symfony\Component\Mailer\Bridge\Amazon\Transport\SesHttpTransport`, `Symfony\Component\Mailer\Bridge\Amazon\Smtp\SesTransport`
+   to `Symfony\Component\Mailer\Bridge\Amazon\Transport\SesSmtpTransport`.
+
 4.3.0
 -----
 
- * added the bridge
+ * Added the bridge